requirements.txt
sentence-transformers
faiss-cpu
numpy
streamlit
openai